editor/libeditor/html/tests/file_select_all_without_body.html

<html>
<head>
<script type="text/javascript">

function is(aLeft, aRight, aMessage)
{
  window.opener.SimpleTest.is(aLeft, aRight, aMessage);
}

function unload()
{
  window.opener.SimpleTest.finish();
}

function boom()
{
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  var root = document.documentElement;
  while(root.firstChild) {
    root.removeChild(root.firstChild);
  }
  root.appendChild(document.createTextNode("Mozilla"));
  root.focus();
  cespan = document.createElementNS("http://www.w3.org/1999/xhtml", "span");
  cespan.setAttributeNS(null, "contenteditable", "true");
  root.appendChild(cespan);
  try {
    document.execCommand("selectAll", false, null);
  } catch(e) { }

  is(window.getSelection().toString(), "Mozilla",
     "The nodes are not selected");

  window.close();
}

</script></head>

<body onload="boom();" onunload="unload();"></body>
</html>
